refine oop structure
