+++ src/main/java/TaskList.java
@@ -1,0 +1,82 @@
+import java.util.ArrayList;
+import java.util.List;
+
+public class TaskList {
+    private static final String indent = "    ";
+    private final List<Task> tasks;
+
+    public TaskList() {
+        this.tasks = new ArrayList<>();
+    }
+
+    public void addTodo(String s) {
+        Todo t = new Todo(s);
+        tasks.add(t);
+        this.printAdd();
+    }
+
+    public void addDdl(String s) {
+        int i = s.indexOf(" /by ");
+        if (i > 0 && i + 5 < s.length()) {
+            Deadline t = new Deadline(s.substring(0, i), s.substring(i + 5));
+            tasks.add(t);
+            this.printAdd();
+        } else {
+            throw new DukeException("The description of a deadline should be \"<task> /by <time>\".");
+        }
+    }
+
+    public void addEvt(String s) {
+        int i = s.indexOf(" /at ");
+        if (i > 0 && i + 5 < s.length()) {
+            Event t = new Event(s.substring(0, i), s.substring(i + 5));
+            tasks.add(t);
+            this.printAdd();
+        } else {
+            throw new DukeException("The description of an event should be \"<task> /at <time>\".");
+        }
+    }
+
+    private void printAdd() {
+        System.out.println(indent + "Got it. I've added this task:");
+        int n = tasks.size();
+        System.out.println(indent + "  " + tasks.get(n - 1));
+        System.out.println(indent + "Now you have " + n + " tasks in the list.");
+    }
+
+    public void list() {
+        if (tasks.size() == 0) {
+            System.out.println(indent + "You don't have tasks listed.");
+        } else {
+            System.out.println(indent + "Here are the tasks in your list:");
+            for (int i = 0; i < tasks.size(); i++) {
+                System.out.print(indent);
+                System.out.print(i + 1);
+                System.out.println("." + tasks.get(i));
+            }
+        }
+    }
+
+    public void mark(int index, boolean done) {
+        if (index >= tasks.size() || index < 0) {
+            throw new DukeException("Please enter a valid index.");
+        }
+        tasks.set(index, tasks.get(index).mark(done));
+        if (done) {
+            System.out.println(indent + "Nice! I've marked this task as done:");
+        } else {
+            System.out.println(indent + "OK, I've marked this task as not done yet:");
+        }
+        System.out.println(indent + "  " + tasks.get(index));
+    }
+
+    public void delete(int index) {
+        if (index >= tasks.size() || index < 0) {
+            throw new DukeException("Please enter a valid index.");
+        }
+        Task t = tasks.remove(index);
+        System.out.println(indent + "Noted. I've removed this task:");
+        System.out.println(indent + "  " + t);
+        System.out.println(indent + "Now you have " + tasks.size() + " tasks in the list.");
+    }
+}
